Move selectOrganizationAndCourse action to own file
