+++ src/actions/index.ts
@@ -4,6 +4,7 @@ export * from "./downloadNewExercisesForCourse";
 export * from "./downloadOrUpdateCourseExercises";
 export * from "./moveExtensionDataPath";
 export * from "./refreshLocalExercises";
+export * from "./selectOrganizationAndCourse";
 export * from "./user";
 export * from "./webview";
 export * from "./workspace";
