+++ src/actions/selectOrganizationAndCourse.ts
@@ -1,0 +1,120 @@
+import { Err, Ok, Result } from "ts-results";
+
+import TemporaryWebview from "../ui/temporaryWebview";
+import { Logger } from "../utils";
+
+import { ActionContext } from "./types";
+
+/**
+ * Creates a new temporary webview where user can select an organization and a course.
+ */
+export async function selectOrganizationAndCourse(
+    actionContext: ActionContext,
+): Promise<Result<{ organization: string; course: number }, Error>> {
+    const { resources, ui } = actionContext;
+
+    const tempView = new TemporaryWebview(resources, ui);
+
+    let organizationSlug: string | undefined;
+    let courseId: number | undefined;
+
+    while (!(organizationSlug && courseId)) {
+        const orgResult = await selectOrganization(actionContext, tempView);
+        if (orgResult.err) {
+            tempView.dispose();
+            return orgResult;
+        }
+        Logger.log(`Organization slug ${orgResult.val} selected`);
+        organizationSlug = orgResult.val;
+        const courseResult = await selectCourse(actionContext, organizationSlug, tempView);
+        if (courseResult.err) {
+            tempView.dispose();
+            return courseResult;
+        }
+        if (courseResult.val.changeOrg) {
+            continue;
+        }
+        courseId = courseResult.val.course;
+    }
+    Logger.log(`Course with id ${courseId} selected`);
+    tempView.dispose();
+    return new Ok({ organization: organizationSlug, course: courseId });
+}
+
+async function selectOrganization(
+    actionContext: ActionContext,
+    webview?: TemporaryWebview,
+): Promise<Result<string, Error>> {
+    const { tmc, resources, ui } = actionContext;
+
+    const result = await tmc.getOrganizations();
+    if (result.err) {
+        return result;
+    }
+    const organizations = result.val.sort((org1, org2) => org1.name.localeCompare(org2.name));
+    const pinned = organizations.filter((organization) => organization.pinned);
+    const data = { organizations, pinned };
+    let slug: string | undefined;
+
+    await new Promise<void>((resolve) => {
+        const temp = webview || new TemporaryWebview(resources, ui);
+        temp.setContent({
+            title: "Select organization",
+            template: { templateName: "organization", ...data },
+            messageHandler: (msg: { type?: string; slug?: string }) => {
+                if (msg.type !== "setOrganization") {
+                    return;
+                }
+                slug = msg.slug;
+                if (!webview) {
+                    temp.dispose();
+                }
+                resolve();
+            },
+        });
+    });
+    if (!slug) {
+        return new Err(new Error("Couldn't get organization"));
+    }
+    return new Ok(slug);
+}
+
+async function selectCourse(
+    actionContext: ActionContext,
+    orgSlug: string,
+    webview?: TemporaryWebview,
+): Promise<Result<{ changeOrg: boolean; course?: number }, Error>> {
+    const { tmc, resources, ui } = actionContext;
+    const result = await tmc.getCourses(orgSlug);
+
+    if (result.err) {
+        return result;
+    }
+    const courses = result.val.sort((course1, course2) => course1.name.localeCompare(course2.name));
+    const organization = (await tmc.getOrganization(orgSlug)).unwrap();
+    const data = { courses, organization };
+    let changeOrg = false;
+    let course: number | undefined;
+
+    await new Promise<void>((resolve) => {
+        const temp = webview || new TemporaryWebview(resources, ui);
+        temp.setContent({
+            title: "Select course",
+            template: { templateName: "course", ...data },
+            messageHandler: (msg: { type?: string; id?: number }) => {
+                if (msg.type === "setCourse") {
+                    course = msg.id;
+                } else if (msg.type === "changeOrg") {
+                    changeOrg = true;
+                } else {
+                    return;
+                }
+                if (!webview) {
+                    temp.dispose();
+                }
+                resolve();
+            },
+        });
+    });
+    return new Ok({ changeOrg, course });
+}
